requirements.txt
fastapi>=0.75.0
ruamel.yaml==0.17.21
ruamel.yaml.clib>=0.2.6
uvicorn==0.17.5
gunicorn
dnspython
python-multipart
PyMongo>=3.12.0
pymongo[srv]
pip==21.3.1
wheel==0.37.1